This in combination with 877ec0c0b0428fffe2cabf6a7c1ad257be7a28c2 fixes #1.
Variables which are not sent uniformly to the bottom model
are not uniformly sent to the atmosphere. In that case they are averaged.

Since commit 877ec0c0b0428fffe2cabf6a7c1ad257be7a28c2 the averaging
is done directly before sending.
Thus it is now removed from flux-calculating routines.

The averaging is also done on the u- and v-grid. For that purpose
the fraction of area variable has to be regridded.
This should be switched on the flux_calculator.nml from now on.